+++ projects/ngx-amap/src/types/class/transfer/amap.transfer-policy.ts
@@ -1,0 +1,8 @@
+export const TransferPolicy = {
+  LEAST_TIME: 0,
+  LEAST_FEE: 1,
+  LEAST_TRANSFER: 2,
+  LEAST_WALK: 3,
+  MOST_COMFORT: 4,
+  NO_SUBWAY: 5
+};