+++ projects/ngx-amap/src/types/interface/transfer/transfer-options.interface.ts
@@ -1,0 +1,15 @@
+import { Map } from '../../class/amap.map';
+
+export interface TransferOptions {
+  city?: string;
+  nightflag?: boolean;
+  policy?: any;
+  cityd?: string;
+  map?: Map;
+  panel?: any;
+  extensions?: string;
+  hideMarkers?: boolean;
+  isOutline?: boolean;
+  outlineColor?: string;
+  autoFitView?: boolean;
+}