+++ src/app/views/amap-driving-demo/simple/simple.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Subscription } from 'rxjs';
 import { AmapDrivingService } from 'ngx-amap';
 
 declare const require: any;
@@ -8,9 +9,11 @@ declare const require: any;
   templateUrl: './simple.component.html',
   styleUrls: ['./simple.component.scss']
 })
-export class SimpleComponent implements OnInit {
+export class SimpleComponent implements OnInit, OnDestroy {
   demo_html = require('!!html-loader!./simple.component.html');
   demo_ts = require('!!raw-loader!./simple.component.ts');
+
+  private _subscription: Subscription;
 
   constructor(private ds: AmapDrivingService) { }
 
@@ -18,25 +21,29 @@ export class SimpleComponent implements OnInit {
   }
 
   async onMapReady(map) {
-    // 构造AMap.Driving插件的Wrapper
+    // 使用of方法构造AMap.Driving插件的Wrapper
     const driving = await this.ds.of({
       map: map,
       panel: 'panel'
     });
 
-    // 使用插件Wrapper搜索路径，这里使用起点和终点的地名
+    // 使用on方法侦听事件：
+    this._subscription = driving.on('complete').subscribe(event => console.log('Driving event: "complete"', event));
+    this._subscription.add(driving.on('error').subscribe(event => console.log('Driving event: "error"', event)));
+
+    // 使用插件Wrapper搜索路径
     const { status, result } = await driving.search([
       { keyword: '北京市地震局(公交站)', city: '北京' },
       { keyword: '亦庄文化园(地铁站)', city: '北京' }
     ]);
-    console.log('driving search');
-    console.log('status:', status);
-    console.log('result:', result);
-
-    if (typeof result === 'string') {
-      return;
-    }
 
     map.setFitView();
   }
+
+  ngOnDestroy() {
+    if (this._subscription) {
+      this._subscription.unsubscribe();
+    }
+  }
+
 }